Refactor capacity pool calculation
Update flow utility function for finding downstream targets. This change
simplifies the logic for calculating the output rate of the last
facility in a capacity pool by directly using the remaining capacity. It
also refactors the `findTargetsWithDownstream` function to more clearly
separate the collection of all targets from the identification of
targets that have other targets in their dependency tree.